requirements.txt
flask
flask-sqlalchemy
python-dotenv
firebase-admin
google-generativeai
markdown
